pyproject.toml
[project]
name = "image-augmentation"
version = "0.1.0"
description = "Image augmentation tool for generating transformed copies of images"
readme = "README.md"
authors = [
    { name = "mithunvoe", email = "kabyasaha1812@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
image-augment = "image_augmentation.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
